+++ vite.app.config.ts
@@ -82,6 +82,7 @@ export default defineConfig({
         'text-edit': resolve(__dirname, 'src/frontend/app/text-edit.html'),
         'text-check': resolve(__dirname, 'src/frontend/app/text-check.html'),
         tags: resolve(__dirname, 'src/frontend/app/tags.html'),
+        feeds: resolve(__dirname, 'src/frontend/app/feeds.html'),
         texts: resolve(__dirname, 'src/frontend/app/texts.html'),
         settings: resolve(__dirname, 'src/frontend/app/settings.html'),
         'text-print': resolve(__dirname, 'src/frontend/app/text-print.html'),